Add files via upload
Setup F12 to turn the prefix on and off for when I'm working with nested tmux sessions.  That way I can run tmux on servers and locally.  On the servers, I keep the config basic and don't use any addons.  May change and I may update the status bar to show the host name but i'm managing that on my local tmux session.